backend/requirements.txt
django
djangorestframework
django-cors-headers
django-environ
mysqlclient
pillow
djangorestframework-simplejwt
whitenoise
django-taggit
django-stubs
django-oauth-toolkit
